requirements.txt
pywebview[qt]
watchdog
beautifulsoup4
